+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.5.0
 - Rename `Avromatic::Model::Decoder` to `MessageDecoder`.
 - Rename `.deserialize` on generated models to `.message_decode`.
-- Add `#avro_encoded_key`, `#avro_encoded_value` and `.decode` methods to
+- Add `#avro_raw_key`, `#avro_raw_value` and `.raw_decode` methods to
   generated models to support encoding and decoding without a schema registry.
 
 ## v0.4.0